pyproject.toml
[project]
name = "keep-mcp"
version = "0.5.3"
description = "MCP Server for Google Keep"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gkeepapi>=0.16.0",
    "mcp[cli]",
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.20.0",
    "toml>=0.10.0",
]
authors = [
    { name = "Jannik Feuerhahn", email = "jannik@feuer.dev" }
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/feuerdev/keep-mcp"
Repository = "https://github.com/feuerdev/keep-mcp"

[project.scripts]
mcp = "server.cli:main"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/server"]